Move Strahlkorper and some related functions.
This is part of issue #3561.

Strahlkorper, StrahlkorperFunctions, and
ChangeCenterOfStrahlkorper have been moved to
NumericalAlgorithms/SphericalHarmonics.
Similar for the associated tests.

There are no code changes except renaming/reordering
includes, and changes to CMakeFiles.
