+++ src/ApparentHorizons/CMakeLists.txt
@@ -8,11 +8,8 @@ add_spectre_library(${LIBRARY})
 spectre_target_sources(
   ${LIBRARY}
   PRIVATE
-  ChangeCenterOfStrahlkorper.cpp
   FastFlow.cpp
   ObjectLabel.cpp
-  Strahlkorper.cpp
-  StrahlkorperFunctions.cpp
   StrahlkorperGr.cpp
   StrahlkorperInDifferentFrame.cpp
   Tags.cpp
@@ -22,14 +19,11 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
-  ChangeCenterOfStrahlkorper.hpp
   ComputeHorizonVolumeQuantities.hpp
   ComputeHorizonVolumeQuantities.tpp
   ComputeItems.hpp
   FastFlow.hpp
   ObjectLabel.hpp
-  Strahlkorper.hpp
-  StrahlkorperFunctions.hpp
   StrahlkorperGr.hpp
   StrahlkorperInDifferentFrame.hpp
   Tags.hpp
