+++ src/NumericalAlgorithms/SphericalHarmonics/CMakeLists.txt
@@ -8,7 +8,10 @@ add_spectre_library(${LIBRARY})
 spectre_target_sources(
   ${LIBRARY}
   PRIVATE
+  ChangeCenterOfStrahlkorper.cpp
   SpherepackIterator.cpp
+  Strahlkorper.cpp
+  StrahlkorperFunctions.cpp
   YlmSpherepack.cpp
   YlmSpherepackHelper.cpp
   )
@@ -17,7 +20,10 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
+  ChangeCenterOfStrahlkorper.hpp
   SpherepackIterator.hpp
+  Strahlkorper.hpp
+  StrahlkorperFunctions.hpp
   YlmSpherepack.hpp
   YlmSpherepackHelper.hpp
   )
