+++ src/NumericalAlgorithms/SphericalHarmonics/Strahlkorper.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "ApparentHorizons/Strahlkorper.hpp"
+#include "NumericalAlgorithms/SphericalHarmonics/Strahlkorper.hpp"
 
 #include <cmath>
 #include <ostream>
