+++ src/NumericalAlgorithms/SphericalHarmonics/Strahlkorper.hpp
