+++ src/NumericalAlgorithms/SphericalHarmonics/StrahlkorperFunctions.hpp
