+++ tests/Unit/ApparentHorizons/CMakeLists.txt
@@ -5,13 +5,10 @@ set(LIBRARY "Test_ApparentHorizons")
 
 set(LIBRARY_SOURCES
   Test_ApparentHorizonFinder.cpp
-  Test_ChangeCenterOfStrahlkorper.cpp
   Test_ComputeHorizonVolumeQuantities.cpp
   Test_ComputeItems.cpp
   Test_FastFlow.cpp
   Test_ObjectLabel.cpp
-  Test_Strahlkorper.cpp
-  Test_StrahlkorperFunctions.cpp
   Test_StrahlkorperGr.cpp
   Test_StrahlkorperInDifferentFrame.cpp
   Test_Tags.cpp
